Prevent more circular deps
